订单页面优化，抽奖流程优化
订单页面优化，抽奖流程优化

+++ ME时代/Classes/Main/HomePage/Controller/Prize/MEPrizeListVC.m
@@ -102,7 +102,7 @@
     kMeWEAKSELF
     cell.tapBlock = ^{
         kMeSTRONGSELF
-        MEJoinPrizeVC *vc = [[MEJoinPrizeVC alloc] initWithActivityId:[NSString stringWithFormat:@"%ld",model.idField]];
+        MEJoinPrizeVC *vc = [[MEJoinPrizeVC alloc] initWithActivityId:[NSString stringWithFormat:@"%ld",(long)model.idField]];
         vc.finishBlock = ^{
             [strongSelf.refresh reload];
         };
